+++ src/strigui/window.clj
@@ -1,41 +1,34 @@
 (ns strigui.window
-  (:require [clojure2d.core :as c2d])
+  (:require [strigui.widget :as wdg]
+            [capra.core :as c])
   (:import [javax.swing JFrame]))
 
-(defn init-window
-  "Creates a window based on the canvas size. Alternatively, also accepts an already
-   existing window."
-  ([^clojure2d.core.Window window]
-   (let [title (:window-name window)
-         window (assoc window :window-name "main-window")]
-     ;; change the title of the underlying frame, so it doesn't mess with the
-     ;; window name used for the clojure2d events
-     (.setTitle (:frame window) title)))
-  ([width height ^String title] (init-window width height title 10 :mid))
-  ([width height ^String title fps quality]
-   (let [canvas (c2d/canvas width height quality)
-         window (c2d/show-window canvas "main-window" width height fps)
-         new-context {:canvas canvas
-                      :window window}]
-     ;; change the title of the underlying frame, so it doesn't mess with the
-     ;; window name used for the clojure2d events
-     (.setTitle (:frame window) title)
-     (.setDefaultCloseOperation (:frame window) (. JFrame EXIT_ON_CLOSE))
-     new-context)))
+(defrecord Window [name context props]
+ wdg/Widget
+  (coord [this _] this)
+  (defaults [this] this)
+  (before-drawing [this] this)
+  (draw [this _] this)
+  (after-drawing [this] this))
 
-(defn close-window
-  [^clojure2d.core.Window window]
-  (c2d/close-window window))
-
-(defn display-info [context text]
-  (let [canvas (:canvas context)
-        height (c2d/height canvas)
-        width (c2d/width canvas)
-        [_ _ _ h] (c2d/with-canvas-> canvas
-                    (c2d/text-bounding-box text))]
-    (c2d/with-canvas-> canvas
-      (c2d/set-color :white)
-      (c2d/rect 0 (- height h) width h)
-      (c2d/set-font-attributes 15)
-      (c2d/set-color :black)
-      (c2d/text (str text) 0 height))))
+(defn window
+  "Initializes a new window or reuses an existing one
+   context - an already existing context (experimental)
+   x - x position on the screen
+   y - y position on the screen
+   width - width of the window
+   height - height of the window
+   title - name displayed in the title bar of the window
+   color - java.awt.Color of the windows background color
+   rendering-hints - map of java.awt.RenderingHints key value combinations to configure the rendering quality
+   of any widget drawn within the window" 
+  ([x y width height title]
+   (window x y width height title (java.awt.Color. 44 44 44) {java.awt.RenderingHints/KEY_ANTIALIASING java.awt.RenderingHints/VALUE_ANTIALIAS_ON}))
+  ([x y width height title color]
+   (window x y width height title color {java.awt.RenderingHints/KEY_ANTIALIASING java.awt.RenderingHints/VALUE_ANTIALIAS_ON
+                                          java.awt.RenderingHints/KEY_RENDERING java.awt.RenderingHints/VALUE_RENDER_SPEED}))
+  ([x y width height title color rendering-hints]
+   (let [context (c/create-window x y width height title (eval color))
+         context (assoc-in context [:canvas :rendering] rendering-hints)
+         context (update context :canvas c/attach-buffered-strategy 2)]
+     (Window. title context {:x x :y y :width width :height height :color {:background color}}))))